Checkpoint (Read Desc)
Creates table information unique to user id for each click.
Next:
When user logs in, retrieve the data from the table corresponding to their user ID. If that doesn't exist, make a table.
Set that data to the used data in the store, such as the table money becoming current money for rolls.
Then, clicking the roll will replace the old with the new.